Refactor codebase for better maintainability and clean code
- Refactor `Controller` by moving actions to `browser_ai/actions/` with categorized modules (navigation, interaction, extraction, utility).
- Refactor `Agent` by extracting media generation logic to `browser_ai/agent/media.py`.
- Refactor `browser_ai_gui` by moving `TaskManager` to `browser_ai_gui/services/task_manager.py`.
- Fix `EventType` usage in `test_extension_server.py`.
- General cleanup of imports and structure.
