+++ browser_ai/actions/__init__.py
@@ -1,0 +1,4 @@
+from .navigation import *
+from .interaction import *
+from .extraction import *
+from .utility import *
